added routes and links
