+++ src/pages/Community.jsx
@@ -1,0 +1,5 @@
+const Community = (props) => {
+    return <h1>This is the Community Page</h1>
+}
+
+export default Community
