+++ src/pages/Home.jsx
@@ -1,0 +1,5 @@
+const Home = (props) => {
+    return <h1>Landing Page</h1>
+}
+
+export default Home
